# doc/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://43.128.109.91:8801",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/collect/messages": {
            "post": {
                "tags": [
                    "\u91c7\u96c6\u63a5\u53e3 (Collect)"
                ],
                "summary": "\u91c7\u96c6\u804a\u5929\u8bb0\u5f55 (WS code=17)",
                "description": "\u5165\u53c2\u53ef\u662f {code:17,data:{list:[...]}} \u6216 {list:[...]} \u6216\u76f4\u63a5 [...] \u6570\u7ec4",
                "operationId": "messages",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/collect/friends": {
            "post": {
                "tags": [
                    "\u91c7\u96c6\u63a5\u53e3 (Collect)"
                ],
                "summary": "\u91c7\u96c6\u8d26\u53f7\u7684\u597d\u53cb\u5217\u8868 (WS code=7)",
                "description": "\u5165\u53c2\u540c\u4e0a, \u6bcf\u6761\u81ea\u5e26 accountId \u8bf4\u660e\u5f52\u5c5e\u54ea\u4e2a\u8d26\u53f7",
                "operationId": "friends",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/collect/accounts": {
            "post": {
                "tags": [
                    "\u91c7\u96c6\u63a5\u53e3 (Collect)"
                ],
                "summary": "\u91c7\u96c6\u8d26\u53f7\u5217\u8868 (WS code=1)",
                "description": "\u5165\u53c2\u53ef\u76f4\u63a5\u662f WS \u8fd4\u56de\u4f53 {code:1,data:[...]} \u6216\u8005\u53ea\u6709 data \u6570\u7ec4",
                "operationId": "accounts",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "JsonNode": {
                "type": "object"
            },
            "RMapStringObject": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    }
}